Increase max_retries
This pull request:

- Increases the maximum number of Boto call retries from 5 to 30
- Switches from an exponential backoff to a linear backoff
- Renames exponential_backoff to _retry (as it's no longer exponential)
- Moves function definition from helpers.py to connection_manager.py
because it's only used by the connection manager
